rectangle f64

+++ crates/crabe_math/src/shape/rectangle.rs
@@ -1,7 +1,7 @@
 use nalgebra::Point2;
 
 pub struct Rectangle {
-    pub width: f32,
-    pub depth: f32,
-    pub left_position: Point2<f32>,
+    pub width: f64,
+    pub depth: f64,
+    pub left_position: Point2<f64>,
 }
